Second session. reviewing application flow demonstrate inheritance via extention in attribute classes.

+++ src/main/java/com/example/demo/repository/CharacterRepository.java
@@ -1,9 +1,10 @@
 package com.example.demo.repository;
 
 import com.example.demo.entity.CharacterEntity;
+import com.example.demo.model.Character;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends JpaRepository<CharacterEntity, Long> {
+public interface CharacterRepository extends JpaRepository<CharacterEntity, Integer> {
 }
